Generalized posits, lns, and multi-dimensional lns (#315)
* bumping SEMVER to v3.63

* engaging Github actions for the new v3.63 branch

* moving Douglas Priest and DoubleDouble to application/numeric

* oracle skeletons for phi, e, and pi

* generating digits for the golden ratio phi

* fix dynamic range bug in quadratic equation

* adding binary string parse and to_hex function to arbitrary precision binary integer

* adding a simplified floating-point engine for Dragon4 and Grisu3 algorithms

* adding multiplication to gfp

* adding a to_hex function to native integer manipulators

* compilation fixes for gcc

* code hygiene

* adding a radix indicator to gfp

* adding cached powers of 10 to gfp

* reverse engineering tests for cached powers of 10 in Grisu paper

* introducing floating-point 2 string class for Grisu algorithms

* adding exponent and significant extractors for IEEE-754 that do not depend on frexp()

* bringing back the q parameter so that we can use f2s as a working floating-point arithmetic

* compilation fix for clang

* adding openbenchmarking.org reference: we should add a mixed-precision benchmark

* WIP: debugging grisu

* troubleshooting operato*() overload not picking up posit specialization

* adding fast bfloat16 implementation

* WIP: streamlining eposit API

* WIP: fixing all the number traits of bfloat16 and disambiguating the origina bfloat16 alias from cfloat

* restructuring the traits as we have generalized them

* adding bfloat16 traits test

* going back to generic size_type in matrix<> BLAS functions

* WIP: adding random test suite for bfloat16

* arithmetic randoms for bfloat16: this is a new lightweight pattern for regression testing

* enabling debug info to troubleshoot CI failures

* fixing clang compilation errors

* WIP: troubleshooting regression failures with gcc11

* adding number system sampling error benchmarks

* compilation fixes for gcc

+++ internal/blocksignificant/arithmetic/addition.cpp
@@ -15,10 +15,10 @@
 // enumerate all addition cases for an blocksignificant<nbits,BlockType> configuration
 template<typename blocksignificantConfiguration>
 int VerifyBlockSignificantAddition(bool reportTestCases) {
-	constexpr size_t nbits = blocksignificantConfiguration::nbits;
+	constexpr unsigned nbits = blocksignificantConfiguration::nbits;
 	using BlockType = typename blocksignificantConfiguration::BlockType;
 
-	constexpr size_t NR_VALUES = (size_t(1) << nbits);
+	constexpr unsigned NR_VALUES = (1u << nbits);
 	using namespace sw::universal;
 	
 //	cout << endl;
